feat: move the main functionality to the 'code' package
