+++ cmd/hexlet-path-size/main.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"os"
 
+	"code"
+
 	cli "github.com/urfave/cli/v3"
 )
 
@@ -20,7 +22,7 @@ func main() {
 
 			path := os.Args[1]
 
-			resStr, err := GetPathSize(path, false, true, false)
+			resStr, err := code.GetPathSize(path, false, true, false)
 
 			if err != nil {
 				return err
@@ -34,48 +36,4 @@ func main() {
 	if err := cmd.Run(context.Background(), os.Args); err != nil {
 		fmt.Println(err.Error())
 	}
-}
-
-func GetPathSize(path string, recursive, human, all bool) (string, error) {
-	size, err := getSize(path, false, false)
-
-	result := fmt.Sprintf("%d\t%s", size, path)
-
-	return result, err
-}
-
-func getSize(path string, recursive, all bool) (int64, error) {
-	f_info, err := os.Lstat(path)
-
-	if err != nil {
-		return 0, err
-	}
-
-	var size int64 = 0
-
-	if f_info.IsDir() {
-		d_entry, err := os.ReadDir(path)
-
-		if err != nil {
-			return 0, err
-		}
-
-		for _, entry := range d_entry {
-			if entry.IsDir() {
-				continue
-			}
-
-			f_info, err := entry.Info()
-
-			if err != nil {
-				return 0, err
-			}
-
-			size += f_info.Size()
-		}
-	} else {
-		size = f_info.Size()
-	}
-
-	return size, nil
 }
