+++ go.mod
@@ -2,4 +2,4 @@ module code
 
 go 1.24.4
 
-require github.com/urfave/cli/v3 v3.3.8 // indirect
+require github.com/urfave/cli/v3 v3.3.8
